# tests/market/contracts/data_plane_openapi.json
{
  "components": {
    "schemas": {
...
      "CompanyFundamentals": {
        "description": "Company details plus key metrics and ratios overview.\n\nAggregate returned by ``GET /api/v1/market/data/fundamentals/{symbol}``:\nthe canonical ``profile``, ``key_metrics`` and ``ratios`` objects, each\nfield-for-field. Profile is required; key metrics and ratios are nullable\nwhen not reported by upstream providers (e.g. ETFs, recent IPOs).",
        "properties": {
          "key_metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeyMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile": {
            "$ref": "#/components/schemas/CompanyProfile"
          },
          "ratios": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialRatios"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "profile"
        ],
        "title": "CompanyFundamentals",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionExpirations": {
        "description": "Available option expiration dates for an underlying symbol.",
        "properties": {
          "expirations": {
            "default": [],
            "items": {
              "format": "date",
              "type": "string"
            },
            "title": "Expirations",
            "type": "array"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          },
          "underlying_symbol": {
            "title": "Underlying Symbol",
            "type": "string"
          }
        },
        "required": [
          "underlying_symbol"
        ],
        "title": "OptionExpirations",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Market Data Plane Contract",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/market/data/fundamentals/{symbol}": {
      "get": {
        "description": "Company details plus key metrics and ratios, served through the cache.\n\nThe response keeps the canonical ``profile``, ``key_metrics`` and\n``ratios`` objects field-for-field; ``exchange`` is forwarded so non-US\nsymbols map to the provider's ticker suffix. Key metrics and ratios are\nnullable when unavailable upstream.",
        "operationId": "market_data_fundamentals_api_v1_market_data_fundamentals__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exchange",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "NYSE",
                    "NASDAQ",
                    "NYSEARCA",
                    "AMEX",
                    "TSX",
                    "LSE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exchange"
            }
          },
          {
            "in": "header",
            "name": "X-Service-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyFundamentals"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Data Fundamentals"
      }
    },
    "/api/v1/market/data/fundamentals/{symbol}/statements": {
      "get": {
        "description": "Statement list for a symbol, served through the cache.\n\n``statement`` and ``period`` are ``Literal``-typed so an invalid value is a\nFastAPI 422; the selected provider read keeps every line item the provider\nreturns (the response is the bare statement list, matching the search\nroute). ``exchange`` is forwarded for non-US ticker mapping.",
        "operationId": "market_data_statements_api_v1_market_data_fundamentals__symbol__statements_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "statement",
            "required": true,
            "schema": {
              "enum": [
                "income",
                "balance",
                "cashflow"
              ],
              "title": "Statement",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "annual",
              "enum": [
                "annual",
                "quarter"
              ],
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "exchange",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "NYSE",
                    "NASDAQ",
                    "NYSEARCA",
                    "AMEX",
                    "TSX",
                    "LSE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exchange"
            }
          },
          {
            "in": "header",
            "name": "X-Service-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "items": {
                        "$ref": "#/components/schemas/IncomeStatement"
                      },
                      "type": "array"
                    },
                    {
                      "items": {
                        "$ref": "#/components/schemas/BalanceSheet"
                      },
                      "type": "array"
                    },
                    {
                      "items": {
                        "$ref": "#/components/schemas/CashFlowStatement"
                      },
                      "type": "array"
                    }
                  ],
                  "title": "Response Market Data Statements Api V1 Market Data Fundamentals  Symbol  Statements Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Data Statements"
      }
    },
    "/api/v1/market/data/options/{symbol}": {
      "get": {
        "description": "Options chain for an underlying.",
        "operationId": "market_data_options_api_v1_market_data_options__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expiry",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Expiry",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "option_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "call",
                    "put"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Option Type"
            }
          },
          {
            "in": "query",
            "name": "strike_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strike Min"
            }
          },
          {
            "in": "query",
            "name": "strike_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strike Max"
            }
          },
          {
            "in": "header",
            "name": "X-Service-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsChain"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Data Options"
      }
    },
    "/api/v1/market/data/options/{symbol}/expirations": {
      "get": {
        "description": "Available option expiration dates for an underlying symbol.",
        "operationId": "market_data_option_expirations_api_v1_market_data_options__symbol__expirations_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Service-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionExpirations"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Data Option Expirations"
      }
    },
    "/api/v1/market/data/prices/{symbol}": {
      "get": {
        "description": "Daily price history for a symbol, served through the endpoint cache.",
        "operationId": "market_data_prices_api_v1_market_data_prices__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exchange",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "NYSE",
                    "NASDAQ",
                    "NYSEARCA",
                    "AMEX",
                    "TSX",
                    "LSE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exchange"
            }
          },
          {
            "in": "header",
            "name": "X-Service-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Data Prices"
      }
    },
...
  }
}